fixes to isoch accounting when FPS less than 1

+++ include/Reporter.h
@@ -274,7 +274,7 @@ struct ShiftCounters {
 };
 
 struct IsochStats {
-    int mFPS; //frames per second
+    double mFPS; //frames per second
     double mMean; //variable bit rate mean
     double mVariance; //vbr variance
     int mJitterBufSize; //Server jitter buffer size, units is frames
